% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_conversion.R
\name{conv_stamp}
\alias{conv_stamp}
\title{Convert timestamp to date time object}
\usage{
conv_stamp(comp_time = NULL, init_stamp = NULL, now_stamp)
}
\arguments{
\item{comp_time}{POSIXlt object}

\item{init_stamp}{state of program timer at comp_time}

\item{now_stamp}{current state of program timer}
}
\value{
POSIXlt with correct date/time
}
\description{
Converts the timestamp on an \*.n38 record to computer time
}
\keyword{internal}
