% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialise.R
\name{em38_pair}
\alias{em38_pair}
\title{Reconcile locations of paired data}
\usage{
em38_pair(decode = NULL, time_filter = NULL)
}
\arguments{
\item{decode}{spatial point dataframe for a survey line produced by
\code{\link{em38_decode}} or \code{\link{em38_from_file}}.}

\item{time_filter}{removes point pairs that are close together but that were
sampled more than n seconds apart.}
}
\value{
An sf data frame with sfc_POINT geometry. WGS84 projection. Output
  locations are averages of input horizontal/vertical paired locations.
}
\description{
Where paired horizontal and vertical readings have been taken during a
'manual' mode survey, the first and second readings at each station should
have the same location. The nature of the device logging generally precludes
this from happening by default, especially with high-frequency GPS recording.
This function reconciles the locations of such paired datasets after they
have been generated using \code{\link{em38_decode}} or
\code{\link{em38_from_file}}.
}
\note{
Input survey should be of survey type 'GPS' and record type 'manual'.
  Both input datasets should ideally have the same number of rows, with row 1
  of horizontal_data paired with row 1 of vertical_data.
}
