% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialise.R
\name{em38_surveyline}
\alias{em38_surveyline}
\title{Process EM38 Survey line}
\usage{
em38_surveyline(survey_line = NULL, hdop_filter = NULL,
  time_filter = NULL, fix_filter = NULL)
}
\arguments{
\item{survey_line}{a decoded survey line from \code{\link{n38_decode}}.}

\item{hdop_filter}{Numeric, discard GPS data where the Horizontal Dilution of
Precision is greater than this number. Defaults to 3 metres. Set to NULL to
keep all readings.}

\item{time_filter}{Numeric, discard readings taken more than \code{n} seconds
after the last acceptable GPS reading. Usually best to set this to 2-3x GPS
aquisition frequency.}

\item{fix_filter}{Select readings with a minimum quality of GPS fix.
  Options are:
 \enumerate{
  \item Autonomous GPS fix.
  \item DGPS fix, using a local DGPS base station or correction service
  such as WAAS or EGNOS.
  \item Pulse per second (PPS) fix.
  \item real-time kinematic (RTK) fix.
  \item RTK float fix.
  \item estimated (dead reckoning)
  \item manual input mode
  \item simulation mode
  \item WAAS fix.
  }
To filter on multiple options, supply a vector e.g. \code{c(2, 9)}.}
}
\value{
If valid GPS data is present, an `sf` data frame with sfc_POINT
  geometry. Otherwise, if valid instrument data is present, a data frame.
  Otherwise, a string explaining failure to process.
}
\description{
This function processes an EM38 survey line into a data frame, which may be
spatial.
}
\note{
Inclusion of comments is yet to be implemented.
}
\examples{
data('n38_demo')
n38_chunks  <- n38_chunk(n38_demo)
n38_decoded <- n38_decode(n38_chunks)
demo_line_1 <- em38_surveyline(n38_decoded[[2]], 3)
demo_line_2 <- em38_surveyline(n38_decoded[[2]], 3, 0.2)

}
