% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_conversion.R
\name{get_cond}
\alias{get_cond}
\title{Signal to conductivity/inphase reading}
\usage{
get_cond(signal = NULL)
}
\arguments{
\item{signal}{Integer.}
}
\value{
Numeric; uncalibrated instrument reading.
}
\description{
This function converts an EM38-MK2 signal recieved on any of Channels 1 to 4. Additional
conversion is required for inphase signals, and further application of a calibration factor is
 required to derive the final conductivity or resistivity value. Do not use this function in
 isolation.
}
\examples{
channel_1_or_3 <- em38:::get_cond(30456)
channel_2 <- em38:::get_cond(30456) * 0.00720475
channel_4 <- em38:::get_cond(30456) * 0.028819

}
\keyword{Internal}
