% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialise.R
\name{get_loc_data}
\alias{get_loc_data}
\title{Translate GPS data}
\usage{
get_loc_data(block = NULL)
}
\arguments{
\item{block}{Data frame holding GPS message data, usually a subset of $location_data in a
decoded n38 object}
}
\value{
data frame with a single row
}
\description{
This function decodes GPS data for use in em38_spatial and returns it along with its timestamp
data.
}
