% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_conversion.R
\name{gga_lat}
\alias{gga_lat}
\title{G*GGA latitude}
\usage{
gga_lat(lat = NULL, dir = NULL)
}
\arguments{
\item{lat}{String containing latitude information. G*GGA strings are
comma-separated, latitude is in the second delimited position}

\item{dir}{String determining whether latitude is N or S of the equator.
G*GGA strings are comma-separated, NS status is in the third delimited
position}
}
\value{
Numeric, latitude in decimal degrees.
}
\description{
This function retrieves latitude position information from a G*GGA string.
}
\note{
Inputting a numeric to lat will give incorrect results for latitude -10
  < x < 10 due to loss of leading zero(s).
}
\examples{
lat <- em38:::gga_lat('2729.10198', 'S')

}
\keyword{internal}
