% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{n38_chunk}
\alias{n38_chunk}
\title{Separate message types}
\usage{
n38_chunk(n38_mat = NULL)
}
\arguments{
\item{n38_mat}{The binary matrix output by  \code{\link{n38_import}}.}
}
\value{
A nested list containing binary matrices encoding a) file header
  information and b) each survey line in the input file. For each survey
  line, a list of matrices with survey line data separated by type - header
  and calibration data, instrument readings, position information, and
  comments.
}
\description{
This function chunks imported n38 binary matrices into a list, grouping file
header and survey line information for further processing.
}
\examples{
data('n38_demo')
n38_chunks <- n38_chunk(n38_demo)
}
