% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{n38_decode}
\alias{n38_decode}
\title{process chunked n38 data}
\usage{
n38_decode(chunks = NULL)
}
\arguments{
\item{chunks}{The nested list of raw matrices output by \code{\link{n38_chunk}}.}
}
\value{
A nested list containing decoded n38 data with the following structure:
  - file header information (list)
  - survey line
    - header information (list)
    - calibration data (data frame)
    - timer data (list)
    - instrument readings (timestamped data frame, calibrated)
    - location data (timestamped data frame of NMEA-0138 message strings)
    - new station data (timestamped data frame)
    - comments (timestamped data frame)
}
\description{
This function decodes n38 data according to the specifications provided by the EM38-MKII manual.
}
\examples{
data('n38_demo')
n38_chunks <- n38_chunk(n38_demo)
n38_decoded <- n38_decode(n38_chunks)
}
