% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{process_cal}
\alias{process_cal}
\title{Process calibration data}
\usage{
process_cal(cal_row = NULL)
}
\arguments{
\item{cal_row}{A matrix with 1 row and 25 columns, produced by \code{\link{n38_import}}.}
}
\value{
A list containing 3 information elements recorded in N38 calibration rows.
}
\description{
This function pulls out calibration data from N38 binaries and dumps it into a list.
 This is an internal function with no wider use case.
}
\examples{
data('n38_demo')
n38_chunked <- n38_chunk(n38_demo)
n38_cal1 <- em38:::process_cal(n38_chunked[['survey_line_1']][['cal_data']][1, ])

}
\keyword{Internal}
