% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{process_comment}
\alias{process_comment}
\title{Process comment data}
\usage{
process_comment(comment = NULL)
}
\arguments{
\item{comment}{A matrix with 1 row and 25 columns, produced by \code{\link{n38_import}}.}
}
\value{
A list containing 2 data elements recorded in N38 comment rows.
}
\description{
This function pulls out comment data from N38 binaries and dumps it into a list.
 This is an internal function with no wider use case.
}
\examples{
\dontrun{
data('n38_demo')
n38_chunked <- n38_chunk(n38_demo)
# no comments in this file, soz. sub in yr own data
n38_c1 <- em38:::process_comment(n38_chunked[['survey_line_1']][['comments']][1, ])
}

}
