% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{process_fheader}
\alias{process_fheader}
\title{Process file header}
\usage{
process_fheader(file_header = NULL)
}
\arguments{
\item{file_header}{A matrix with 2 rows and 25 columns, produced by \code{\link{n38_import}}.}
}
\value{
A list containing 9 information elements recorded in N38 file headers.
}
\description{
This function pulls out file header data from N38 binaries and dumps them in a list. This is an
internal function with no wider use case.
}
\examples{
data('n38_demo')
n38_chunked <- n38_chunk(n38_demo)
n38_fh <- em38:::process_fheader(n38_chunked[['file_header']])

}
\keyword{Internal}
