% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_gga}
\alias{process_gga}
\title{Process NMEA-0183 G*GGA messages}
\usage{
process_gga(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 G*GGA structure.}
}
\value{
A list containing the data elements recorded in NMEA-0183 G*GGA data
  chunks. Elements are given appropriate R data types.
}
\description{
This function pulls out position fix data from NMEA-0183 G*GGA strings and
dumps it into a list.
}
\examples{
# first GPGGA msg from data('n38_demo')
msg_1   <- "$GPGGA,015808.00,2726.53758,S,15126.05255,E,1,08,1.0,365.1,M,39.5,M,,*79"
gpgga_1 <- em38:::process_gga(string = msg_1)
}
