% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_gpgga}
\alias{process_gpgga}
\title{Process NMEA-0183 GPGGA messages}
\usage{
process_gpgga(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 GPGGA structure.}
}
\value{
A list containing 15 data elements recorded in NMEA-0183 GPGGA data chunks. Elements
are given appropriate data types. NB UTC time is returned as POSIXlt, so Sys.date() comes along
for the ride.
}
\description{
This function pulls out position fix data from NMEA-0183 GPGGA strings and dumps it into a list.
}
\examples{
\dontrun{
data('n38_demo')
cal <- process_gpgga(n38_demo[??,])}
}
