% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_gpgsv}
\alias{process_gpgsv}
\title{Process NMEA-0183 GPGSV messages}
\usage{
process_gpgsv(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 GPGSV structure.}
}
\value{
A list containing 15 data elements recorded in NMEA-0183 GPGSV data chunks. Elements
are given appropriate data types.
Note that depending on the number of satellites in view, up to three of these messages may
exist for every GPS reading. Up to four satellites are reported on per string.
Note that SNR is receiver-dependant and should only be considered relative to other readings
 in the same dataset.
}
\description{
This function pulls out satellites in view data from NMEA-0183 GPGSV strings
and dumps it into a list.
}
\examples{
\dontrun{
data('n38_demo')
cal <- process_gpgsv(n38_demo[??,])}

}
