% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_gprmc}
\alias{process_gprmc}
\title{Process NMEA-0183 GPRMC messages}
\usage{
process_gprmc(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 GPRMC structure.}
}
\value{
A list containing 9 data elements recorded in NMEA-0183 GPRMC data chunks. Elements
are given appropriate data types.
}
\description{
This function pulls out reccommended minimum sentence data from NMEA-0183 GPRMC strings
and dumps it into a list.
}
\examples{
\dontrun{
data('n38_demo')
cal <- process_gprmc(n38_demo[??,])}

}
