% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_gpvtg}
\alias{process_gpvtg}
\title{Process NMEA-0183 GPVTG messages}
\usage{
process_gpvtg(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 GPVTG structure.}
}
\value{
A list containing 6 data elements recorded in NMEA-0183 GPVTG data
  chunks. Elements are given appropriate data types.
}
\description{
This function pulls out track made good and speed over ground data from
NMEA-0183 GPVTG strings and dumps it into a list.
}
\examples{
# first GPGVTG msg from data('n38_demo')
msg_1   <- "$GPVTG,208.02,T,,M,0.32,N,0.59,K,A*38"
gpvtg_1 <- em38:::process_gpvtg(string = msg_1)
}
