% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_gsa}
\alias{process_gsa}
\title{Process NMEA-0183 G*GSA messages}
\usage{
process_gsa(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 G*GSA structure.}
}
\value{
A list containing n data elements recorded in NMEA-0183 G*GSA data
  chunks. Elements are given appropriate data types.
}
\description{
This function pulls out GPS DOP and active satellites data from NMEA-0183
G*GSA strings and dumps it into a list.
}
\examples{
# first GPGSA msg from data('n38_demo')
msg_1   <- "$GPGSA,M,3,05,10,15,16,20,21,26,29,,,,,1.6,1.0,1.2*32"
gpgsa_1 <- em38:::process_gsa(string = msg_1)

}
