% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{process_reading}
\alias{process_reading}
\title{Process instrument reading data}
\usage{
process_reading(reading = NULL)
}
\arguments{
\item{reading}{A matrix with 1 row and 25 columns, produced by \code{\link{n38_import}}.}
}
\value{
A list containing 10 data elements recorded in N38 instrument reading rows.
}
\description{
This function pulls out instrument reading data from N38 binaries and dumps it into a list. This
is an internal function with no wider use case.
}
\examples{
data('n38_demo')
n38_chunked <- n38_chunk(n38_demo)
n38_r1 <- em38:::process_reading(n38_chunked[['survey_line_1']][['reading_data']][1, ])

}
\keyword{Internal}
