% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{process_rmc}
\alias{process_rmc}
\title{Process NMEA-0183 G*RMC messages}
\usage{
process_rmc(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 G*RMC structure.}
}
\value{
A list containing 9 data elements recorded in NMEA-0183 G*RMC data
  chunks. Elements are given appropriate data types.
}
\description{
This function pulls out reccommended minimum sentence data from NMEA-0183
G*RMC strings and dumps it into a list.
}
\examples{
# first GPRMC msg from data('n38_demo')
msg_1   <- "$GPRMC,015808.00,A,2726.53758,S,15126.05255,E,0.32,208.02,160318,,,A*48"
gprmc_1 <- em38:::process_rmc(string = msg_1)

}
