% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{process_slheader}
\alias{process_slheader}
\title{Process survey line header}
\usage{
process_slheader(survline_header = NULL)
}
\arguments{
\item{survline_header}{A matrix with 4 rows and 25 columns, produced by
\code{\link{n38_import}}.}
}
\value{
A list containing 5 information elements recorded in N38 survey line
  headers.
}
\description{
This function pulls out survey line header data from *.N38 binaries and dumps
them into a list. This is an internal function with no wider use case.
}
\examples{
data('n38_demo')
n38_chunked <- n38_chunk(n38_demo)
n38_slh <- em38:::process_slheader(n38_chunked[['survey_line_1']][['sl_header']])

}
\keyword{internal}
