% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{process_timer}
\alias{process_timer}
\title{Process timer relation data}
\usage{
process_timer(timer_rel = NULL)
}
\arguments{
\item{timer_rel}{A matrix with 1 row and 25 columns, produced by \code{\link{n38_import}}.}
}
\value{
A list containing 2 information elements recorded in N38 timer relation rows. Note that
time will be returned in the local timezone.
}
\description{
This function pulls out timer relation data from N38 binaries and dumps it into a list.
 This is an internal function with no wider use case.
}
\examples{
data('n38_demo')
n38_chunked <- n38_chunk(n38_demo)
n38_t <- em38:::process_timer(n38_chunked[['survey_line_1']][['timer_data']])

}
\keyword{Internal}
