% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_conversion.R
\name{get_temp}
\alias{get_temp}
\title{Signal to temperature reading}
\usage{
get_temp(signal = NULL)
}
\arguments{
\item{signal}{Integer.}
}
\value{
Temperature in degrees C
}
\description{
This function calculates temperature from an EM38-MK2 signal recieved on
Channel 5 or 6.
}
\note{
The device's effective reporting resolution is ~0.32 degrees C.
}
\examples{
channel_5 <- em38:::get_temp(227)

}
\keyword{Internal}
