% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{n38_import}
\alias{n38_import}
\title{Import EM38 data}
\usage{
n38_import(path = NULL, type = NULL)
}
\arguments{
\item{path}{A file path pointing to a valid file, produced by a Geonics
EM38-MK2 conductivity sensor connected to an Allegra CX or Archer
datalogger (N38) or connected through Geomar TackMaker software (382)
(and optionally, a GPS device).}

\item{type}{A file extension, one of either N38 (Geonics) or N382 (Geomar)}
}
\value{
A matrix with n rows and 25 columns, containing raw bytes.
}
\description{
This function reads N38 binary files into R and does some minimal
pre-processing.
}
\examples{
n38_mat <-
  n38_import(system.file("extdata", "em38_demo.N38", package = "em38"), type = "N38")
n38_mat <-
  n38_import(system.file("extdata", "em38_demo.382", package = "em38"), type = "N382")
}
