% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_data.R
\name{nmea_check}
\alias{nmea_check}
\title{Calculate NMEA-0183 XOR checksum}
\usage{
nmea_check(string = NULL)
}
\arguments{
\item{string}{A string with valid NMEA-0183 G*GGA structure.}
}
\value{
Logical; TRUE if checksum is correct.
}
\description{
This function calculates the checksum for NMEA-0183 G*GGA strings.
}
\examples{
# first G*GGA msg from data('n38_demo')
msg_1 <- "$GPGGA,015808.00,2726.53758,S,15126.05255,E,1,08,1.0,365.1,M,39.5,M,,*79"
msg_2 <- "$GPG5808.00,2726.53758,S,15126.05255,E,1,08,1.0,365.1,M,39.5,M,,*79"
msg_3<- "$GNGGA,201703.40,4952.9329190,N,10012.1760775,W,4,24,0.7,364.311,M,-23.378,M,1.4,0000*6A"
chk_1 <- em38:::nmea_check(string = msg_1)
chk_2 <- em38:::nmea_check(string = msg_2)
chk_3 <- nav38:::nmea_check(string = msg_3)

}
